[NMPC][Planning][Refactor]: refactor mpc_pt_pub_node.py that it receives the robot_name as an argument and is universal for all robots.
