+++ aerial_robot_planning/CMakeLists.txt
@@ -156,9 +156,7 @@ include_directories(
 ## Mark executable scripts (Python etc.) for installation
 ## in contrast to setup.py, you can choose the destination
  catkin_install_python(PROGRAMS
-   scripts/beetle_mpc_pt_pub_node.py
-   scripts/tri_mpc_pt_pub_node.py
-   scripts/bi_mpc_pt_pub_node.py
+   scripts/mpc_pt_pub_node.py
    DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
  )
